Remove api tests
